requirements.txt
# Core dependencies
python-telegram-bot==20.7
vkbottle==4.3.12
python-dotenv==1.0.0

# Scheduling and database
apscheduler==3.10.4
sqlalchemy==2.0.23
alembic==1.13.1  # Database migrations
psycopg2-binary==2.9.9  # PostgreSQL driver

# Timezone support
pytz==2023.3.post1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-docx==1.1.0  # For PDF export later

# Optional (для будущих фаз)
# pillow==10.1.0  # Image processing (Фаза 2)
# openai==1.3.0  # AI features (Фаза 3)
# anthropic==0.7.0  # AI features alternative (Фаза 3)
